created new directory history_catalog and functions related to managing history file moved to new grassdb/history.py
