+++ gui/wxpython/lmgr/frame.py
@@ -62,7 +62,7 @@ from gui_core.dialogs import (
     MapLayersDialog,
     QuitDialog,
 )
-from gui_core.menu import SearchModuleWindow, HistoryModuleWindow
+from gui_core.menu import SearchModuleWindow
 from gui_core.menu import Menu as GMenu
 from core.debug import Debug
 from lmgr.toolbars import LMWorkspaceToolbar, LMToolsToolbar
@@ -73,6 +73,7 @@ from lmgr.pyshell import PyShellWindow
 from lmgr.giface import LayerManagerGrassInterface
 from mapdisp.frame import MapDisplay
 from datacatalog.catalog import DataCatalog
+from history_catalog.catalog import HistoryCatalog
 from gui_core.forms import GUI
 from gui_core.wrap import Menu, TextEntryDialog
 from startup.guiutils import (
@@ -164,7 +165,7 @@ class GMFrame(wx.Frame):
         self._createDisplay(self.notebook)
         self._createSearchModule(self.notebook)
         self._createConsole(self.notebook)
-        self._createHistoryModule(self.notebook)
+        self._createHistoryCatalog(self.notebook)
         self._createPythonShell(self.notebook)
         self._addPagesToNotebook()
         self.toolbars = {
@@ -446,10 +447,10 @@ class GMFrame(wx.Frame):
 
         self._setCopyingOfSelectedText()
 
-    def _createHistoryModule(self, parent):
-        """Initialize Module history widget"""
+    def _createHistoryCatalog(self, parent):
+        """Initialize history catalog widget"""
         if not UserSettings.Get(group="manager", key="hideTabs", subkey="history"):
-            self.history = HistoryModuleWindow(parent=parent, giface=self._giface)
+            self.history = HistoryCatalog(parent=parent, giface=self._giface)
             self.history.showNotification.connect(
                 lambda message: self.SetStatusText(message)
             )
