+++ gui/wxpython/main_window/frame.py
@@ -69,7 +69,7 @@ from gui_core.dialogs import (
     MapLayersDialog,
     QuitDialog,
 )
-from gui_core.menu import SearchModuleWindow, HistoryModuleWindow
+from gui_core.menu import SearchModuleWindow
 from gui_core.menu import Menu as GMenu
 from core.debug import Debug
 from lmgr.toolbars import LMWorkspaceToolbar, LMToolsToolbar
@@ -80,6 +80,7 @@ from lmgr.pyshell import PyShellWindow
 from lmgr.giface import LayerManagerGrassInterface
 from mapdisp.frame import MapPanel
 from datacatalog.catalog import DataCatalog
+from history_catalog.catalog import HistoryCatalog
 from gui_core.forms import GUI
 from gui_core.wrap import Menu, TextEntryDialog, SimpleTabArt
 from startup.guiutils import (
@@ -387,10 +388,10 @@ class GMFrame(wx.Frame):
 
         self._setCopyingOfSelectedText()
 
-    def _createHistoryModule(self, parent):
-        """Initialize Module history widget"""
+    def _createHistoryCatalog(self, parent):
+        """Initialize history catalog widget"""
         if not UserSettings.Get(group="manager", key="hideTabs", subkey="history"):
-            self.history = HistoryModuleWindow(parent=parent, giface=self._giface)
+            self.history = HistoryCatalog(parent=parent, giface=self._giface)
             self.history.showNotification.connect(
                 lambda message: self.SetStatusText(message)
             )
@@ -568,7 +569,7 @@ class GMFrame(wx.Frame):
         self._createDisplay(parent=self)
         self._createSearchModule(parent=self)
         self._createConsole(parent=self)
-        self._createHistoryModule(parent=self)
+        self._createHistoryCatalog(parent=self)
         self._createPythonShell(parent=self)
         self.toolbars = {
             "workspace": LMWorkspaceToolbar(parent=self),
