+++ python/grass/grassdb/checks.py
@@ -139,21 +139,6 @@ def get_mapset_owner(mapset_path):
         return None
 
 
-def get_mapset_history_path(mapset_path):
-    """Returns path to the mapset history file or None if mapset has no history."""
-    history_file_name = ".wxgui_history"
-    history_file = os.path.join(mapset_path, history_file_name)
-    return history_file if os.path.exists(history_file) else None
-
-
-def get_current_mapset_history_path():
-    """Returns path to the current mapset history file
-    or None if current mapset has no history."""
-    genv = gisenv()
-    mapset_path = Path(genv["GISDBASE"]) / genv["LOCATION_NAME"] / genv["MAPSET"]
-    return get_mapset_history_path(mapset_path)
-
-
 def is_fallback_session():
     """Checks if a user encounters a fallback GRASS session.
